Split up cli_schema into various smaller ones
